refactor 	- 불필요한 코드 삭제 	- 포인트 생성 역할 포인트 도메인으로 위임
