+++ src/main/java/step2/model/Point.java
@@ -9,12 +9,34 @@ public class Point {
         this.direction = Direction.findDirection(left, current);
     }
 
+
     public boolean getCurrent() {
         return direction.getCurrent();
     }
 
     public int changeIndex(int index) {
         return index + direction.getMove();
+    }
+    public static Point first(ValueStrategy valueStrategy) {
+        return new Point(false, valueStrategy.generateValue());
+    }
+
+    public Point next(ValueStrategy valueStrategy) {
+        return new Point(this.getCurrent(), generateCurrentValue(valueStrategy, this.getCurrent()));
+    }
+
+    public Point last() {
+        return new Point(this.direction.getCurrent(), false);
+    }
+
+    private static boolean generateCurrentValue(ValueStrategy valueStrategy, boolean beforeCurrent) {
+        boolean current = valueStrategy.generateValue();
+
+        if (beforeCurrent) {
+            current = false;
+        }
+
+        return current;
     }
 
     @Override
@@ -29,4 +51,6 @@ public class Point {
     public int hashCode() {
         return Objects.hash(direction);
     }
+
+
 }
